Move controller to admin folder
